requirements.txt
b3j0f.conf==0.3.18
link.middleware==0.10
link.dbrequest==0.8
six==1.10.0
ipython==4.2.0
